Do not install hypothesis in the wheel-build test env
The wheel-build test run already skips the property-based quoting tests
via -m "not hypothesis", but hypothesis was still installed through
requirements/test-cibuildwheel.txt and imported at collection time by
tests/test_quoting.py, so it was built from source on every arch. On
arches without a prebuilt wheel (armv7l musllinux) that source build
pulls in a Rust toolchain and failed the wheel jobs.

Move the property-based tests into tests/test_quoting_hypothesis.py
behind an importorskip so the module is skipped when hypothesis is
absent, and move the hypothesis requirement from test-cibuildwheel.txt
to test.txt so the regular test run still installs it while the wheel
build does not. This mirrors how pydantic is already scoped to the
regular run only.
