+++ CHANGES/1804.packaging.rst
@@ -1,0 +1,5 @@
+Stopped installing ``hypothesis`` in the wheel-build test environment. The
+property-based quoting tests that need it are skipped there already, and
+building it from source on architectures without a prebuilt wheel (such as
+``armv7l`` musllinux, where the build pulls in a Rust toolchain) was failing
+the wheel jobs -- by :user:`bdraco`.
