+++ requirements/test.txt
@@ -1,4 +1,10 @@
 -r test-cibuildwheel.txt
+# hypothesis is only needed for the property-based tests in
+# tests/test_quoting_hypothesis.py, which the wheel-build test run skips
+# (via -m "not hypothesis" plus an importorskip). Keeping it out of
+# test-cibuildwheel.txt avoids building it from source on arches without a
+# prebuilt wheel (e.g. armv7l musllinux, where the source build needs Rust).
+hypothesis>=6.157.0
 # pydantic-core has no binary wheels for freethreading mode,
 # let's skip testing it during wheels building.
 # The regular test run is sufficient for testing pydantic integration.
