+++ tests/test_quoting.py
@@ -1,6 +1,4 @@
 import pytest
-from hypothesis import assume, example, given, note
-from hypothesis import strategies as st
 
 import yarl
 from yarl._quoting import NO_EXTENSIONS, _Quoter, _Unquoter
@@ -558,79 +556,3 @@ def test_quoter_path_with_plus(quoter: type[_Quoter]) -> None:
 def test_unquoter_path_with_plus(unquoter: type[_Unquoter]) -> None:
     s = "/test/x+y%2Bz/:+%2B/"
     assert "/test/x+y+z/:++/" == unquoter(unsafe="+")(s)
-
-
-@given(safe=st.text(), protected=st.text(), qs=st.booleans(), requote=st.booleans())
-def test_fuzz__PyQuoter(safe: str, protected: str, qs: bool, requote: bool) -> None:  # type: ignore[misc]
-    """Verify that _PyQuoter can be instantiated with any valid arguments."""
-    _PyQuoter(safe=safe, protected=protected, qs=qs, requote=requote)
-
-
-@given(ignore=st.text(), unsafe=st.text(), qs=st.booleans())
-def test_fuzz__PyUnquoter(ignore: str, unsafe: str, qs: bool) -> None:  # type: ignore[misc]
-    """Verify that _PyUnquoter can be instantiated with any valid arguments."""
-    _PyUnquoter(ignore=ignore, unsafe=unsafe, qs=qs)
-
-
-@example(text_input="0")
-@given(
-    text_input=st.text(
-        alphabet=st.characters(max_codepoint=127, blacklist_characters="%")
-    ),
-)
-@pytest.mark.parametrize("quoter", quoters, ids=quoter_ids)
-@pytest.mark.parametrize("unquoter", unquoters, ids=unquoter_ids)
-def test_quote_unquote_parameter(  # type: ignore[misc]
-    quoter: type[_PyQuoter],
-    unquoter: type[_PyUnquoter],
-    text_input: str,
-) -> None:
-    quote = quoter()
-    unquote = unquoter()
-    text_quoted = quote(text_input)
-    note(f"text_quoted={text_quoted!r}")
-    text_output = unquote(text_quoted)
-    assert text_input == text_output
-
-
-@example(text_input="0")
-@given(
-    text_input=st.text(
-        alphabet=st.characters(max_codepoint=127, blacklist_characters="%")
-    ),
-)
-@pytest.mark.parametrize("quoter", quoters, ids=quoter_ids)
-@pytest.mark.parametrize("unquoter", unquoters, ids=unquoter_ids)
-def test_quote_unquote_parameter_requote(  # type: ignore[misc]
-    quoter: type[_PyQuoter],
-    unquoter: type[_PyUnquoter],
-    text_input: str,
-) -> None:
-    quote = quoter(requote=True)
-    unquote = unquoter()
-    text_quoted = quote(text_input)
-    note(f"text_quoted={text_quoted!r}")
-    text_output = unquote(text_quoted)
-    assert text_input == text_output
-
-
-@example(text_input="0")
-@given(
-    text_input=st.text(
-        alphabet=st.characters(max_codepoint=127, blacklist_characters="%")
-    ),
-)
-@pytest.mark.parametrize("quoter", quoters, ids=quoter_ids)
-@pytest.mark.parametrize("unquoter", unquoters, ids=unquoter_ids)
-def test_quote_unquote_parameter_path_safe(  # type: ignore[misc]
-    quoter: type[_PyQuoter],
-    unquoter: type[_PyUnquoter],
-    text_input: str,
-) -> None:
-    quote = quoter()
-    unquote = unquoter(ignore="/%", unsafe="+")
-    assume("+" not in text_input and "/" not in text_input)
-    text_quoted = quote(text_input)
-    note(f"text_quoted={text_quoted!r}")
-    text_output = unquote(text_quoted)
-    assert text_input == text_output
